refactor: #166 kakao/apple과의 communication에서 사용하는 dto의 path, name 변경
